requirements.txt
boto3
Flask
Flask-SocketIO
mysql-connector
flask_cors
requests
